+++ backend/fittoring/src/main/java/fittoring/mentoring/business/service/MentoringReservationService.java
@@ -11,15 +11,12 @@ import org.springframework.stereotype.Service;
 public class MentoringReservationService {
 
     private final ReservationService reservationService;
+    private final SmsMessageFormatter smsMessageFormatter;
     private final SmsRestClientService smsRestClientService;
 
     public ReservationCreateResponse reserveMentoring(ReservationCreateDto dto) {
         ReservationCreateResponse response = reservationService.createReservation(dto);
-        String smsMessage = reservationService.createSmsReservationMessage(new SmsReservationMessageDto(
-                dto.menteeName(),
-                dto.menteePhone(),
-                dto.content()
-        ));
+        String smsMessage = smsMessageFormatter.createSmsReservationMessage(SmsReservationMessageDto.of(dto));
         smsRestClientService.sendSms(response.mentorPhone(), smsMessage);
         return response;
     }
