+++ backend/fittoring/src/main/java/fittoring/mentoring/business/service/ReservationService.java
@@ -31,16 +31,4 @@ public class ReservationService {
         ));
         return ReservationCreateResponse.from(savedReservation);
     }
-
-    public String createSmsReservationMessage(SmsReservationMessageDto response) {
-        return String.format("""
-                        멘토링 신청자: %s\n
-                        신청자 전화번호: %s\n
-                        상담 내용: %s
-                        """,
-                response.menteeName(),
-                response.menteePhone(),
-                response.content()
-        );
-    }
 }
