+++ backend/fittoring/src/main/java/fittoring/mentoring/business/service/dto/SmsReservationMessageDto.java
@@ -5,4 +5,12 @@ public record SmsReservationMessageDto(
         String menteePhone,
         String content
 ) {
+
+    public static SmsReservationMessageDto of(ReservationCreateDto dto) {
+        return new SmsReservationMessageDto(
+                dto.menteeName(),
+                dto.menteePhone(),
+                dto.content()
+        );
+    }
 }
